progress

+++ fuse_detect.py
@@ -7,6 +7,7 @@ import cv2
 import torch
 import torch.backends.cudnn as cudnn
 from numpy import random
+import numpy as np
 
 
 from models.experimental import attempt_load
@@ -130,8 +131,8 @@ def detect(save_img=True): # changed default
 
                     if save_img or view_img: # save_img or view_img:  # Add bbox to image
                         label = f'{names[int(cls)]} {conf:.2f}'
-                        plot_one_box(xyxy, img_rgb, label=label, color=colors[int(cls)], line_thickness=1)
-                        plot_one_box(xyxy, img_th, label=label, color=colors[int(cls)], line_thickness=1)
+                        plot_one_box(xyxy, np.float32(img_rgb), label=label, color=colors[int(cls)], line_thickness=1)
+                        plot_one_box(xyxy, np.float32(img_th), label=label, color=colors[int(cls)], line_thickness=1)
 
             # Print time (inference + NMS)
             print(f'{s}Done. ({(1E3 * (t2 - t1)):.1f}ms) Inference, ({(1E3 * (t3 - t2)):.1f}ms) NMS')
